mist2 core: Add indices functions.
Function 'i' generate number of current join iteration. The first iteration corresponds to "1".
Function 'i0' is same as 'i' but first iteration corresponds to "0".

More detailed description and usage example - in README.

+++ snippets/mist2/src/mist2/core/mist_param_set.cpp
@@ -504,7 +504,7 @@ void ParamSetSlice::nextSet(void)
         if(iter->slice->isSetLast())
         {
             /* Set of subslice is the last one.*/
-            if((iter->index + 1) < iter->values->size())
+            if((iter->index + 1) < (int)iter->values->size())
             {
                 /* But subslice is not last in array. Advance it. */
                 ++iter->index;
@@ -586,7 +586,7 @@ void ParamSetSlice::reset(const ParamSet* value)
 void ParamSetSlice::updateSetLast(const Subslice& subslice)
 {
     isSetLastFlag = isSetLastFlag
-        && ((subslice.index + 1) == subslice.values->size())
+        && ((subslice.index + 1) == (int)subslice.values->size())
         && subslice.slice->isSetLast();
 }
